Adding lint, prettier and coverage

+++ contracts/interfaces/IExchanger.sol
@@ -1,15 +1,16 @@
 pragma solidity ^0.6.2;
 
 interface IExchanger {
-
     function settle(address from, bytes32 currencyKey)
         external
         returns (
-            uint reclaimed,
-            uint refunded,
-            uint numEntries
+            uint256 reclaimed,
+            uint256 refunded,
+            uint256 numEntries
         );
 
-    function maxSecsLeftInWaitingPeriod(address account, bytes32 currencyKey) external view returns (uint);
-
+    function maxSecsLeftInWaitingPeriod(address account, bytes32 currencyKey)
+        external
+        view
+        returns (uint256);
 }
